+++ app/src/main/AndroidManifest.xml
@@ -15,6 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TP3_PARCIAL_BELGRANO_GRUPO4"
         tools:targetApi="31">
+
         <activity
             android:name=".ui.views.MainActivity"
             android:exported="true">
