+++ app/src/main/java/com/example/tp3_parcial_belgrano_grupo4/ui/views/MainActivity.kt
@@ -1,83 +1,7 @@
 package com.example.tp3_parcial_belgrano_grupo4.ui.views
-/*
-import android.os.Bundle
-import com.google.android.material.appbar.MaterialToolbar
-import androidx.appcompat.app.AppCompatActivity
-import androidx.drawerlayout.widget.DrawerLayout
-import androidx.core.view.GravityCompat
-import androidx.fragment.app.Fragment
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.AppBarConfiguration
-import com.example.tp3_parcial_belgrano_grupo4.R
-import com.google.android.material.navigation.NavigationView
-import androidx.navigation.ui.setupWithNavController
-import com.example.tp3_parcial_belgrano_grupo4.ui.views.profile.ProfileFragment
-import com.example.tp3_parcial_belgrano_grupo4.ui.views.setting.SettingFragment
 
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var toolbar: MaterialToolbar
-    private lateinit var drawerLayout: DrawerLayout
-    private lateinit var navigationView: NavigationView
-    private lateinit var navController: NavController
-    private lateinit var appBarConfiguration: AppBarConfiguration
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        toolbar = findViewById(R.id.topAppbar)
-        drawerLayout = findViewById(R.id.drawer_layout)
-        navigationView = findViewById(R.id.navigation_view)
-
-        val navHostFragment =
-            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-        navController = navHostFragment.navController
-
-        val appBarConfiguration = AppBarConfiguration(navController.graph, drawerLayout)
-        toolbar.setupWithNavController(navController, appBarConfiguration)
-
-
-        navigationView.setNavigationItemSelectedListener { menuItem ->
-            val id = menuItem.itemId
-            drawerLayout.closeDrawer(GravityCompat.START)
-
-            when (id) {
-
-                R.id.nav_profile -> {
-                    //navController.navigate(R.id.nav_profile2)
-                   replaceFragment(ProfileFragment())
-                }
-
-                R.id.nav_setting -> {
-
-                    //navController.navigate(R.id.nav_setting2)
-                    replaceFragment(SettingFragment())
-                }
-
-                else -> return@setNavigationItemSelectedListener true
-            }
-
-            true
-        }
-    }
-
-    private fun replaceFragment(fragment: Fragment) {
-        val fragmentManager = supportFragmentManager
-        val fragmentTransaction = fragmentManager.beginTransaction()
-        fragmentTransaction.replace(R.id.drawer_layout, fragment)
-        fragmentTransaction.commit()
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return navController.navigateUp() || super.onSupportNavigateUp()
-    }
-
-
-}
-*/
+import android.graphics.Color
+import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.core.view.GravityCompat
@@ -98,11 +22,13 @@ class MainActivity : AppCompatActivity(){
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
+
         drawerLayout = findViewById(R.id.drawer_layout)
         navigationView = findViewById(R.id.navigation_view)
         navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
 
         setupDrawerLayout()
+
     }
     private fun setupDrawerLayout(){
         val navController = navHostFragment.navController
@@ -110,9 +36,12 @@ class MainActivity : AppCompatActivity(){
 
         NavigationUI.setupActionBarWithNavController(this, navController,drawerLayout)
 
-        navController.addOnDestinationChangedListener{ _, _, _ ->
-            // Aca le digo que quiero que mi icono izquierdo de la appbar sea el del drawer
-            supportActionBar?.setHomeAsUpIndicator(R.drawable.menu_24)
+        navController.addOnDestinationChangedListener { _, destination, _ ->
+            if (destination.id == R.id.nav_home) {
+                supportActionBar?.setHomeAsUpIndicator(R.drawable.menu_24)
+            } else {
+                supportActionBar?.setDisplayHomeAsUpEnabled(true)
+            }
         }
     }
     override fun onSupportNavigateUp(): Boolean {
