+++ app/src/main/java/com/example/tp3_parcial_belgrano_grupo4/ui/views/fragments/HomeFragment.kt
@@ -1,37 +1,34 @@
-package com.example.tp3_parcial_belgrano_grupo4.ui.views.profile
+package com.example.tp3_parcial_belgrano_grupo4.ui.views.fragments
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
-import android.view.Menu
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import com.example.tp3_parcial_belgrano_grupo4.R
 import com.example.tp3_parcial_belgrano_grupo4.databinding.FragmentHomeBinding
-import com.example.tp3_parcial_belgrano_grupo4.databinding.FragmentProfileBinding
-import com.example.tp3_parcial_belgrano_grupo4.ui.views.home.HomeViewModel
+import com.example.tp3_parcial_belgrano_grupo4.ui.views.viewmodels.HomeViewModel
 
-class ProfileFragment : Fragment() {
+class HomeFragment : Fragment() {
 
-    private var _binding: FragmentProfileBinding? = null
+    private var _binding: FragmentHomeBinding? = null
 
     private val binding get() = _binding!!
-
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val profileViewModel =
-            ViewModelProvider(this).get(ProfileViewModel:: class.java)
 
-        _binding = FragmentProfileBinding.inflate(inflater, container, false)
+        val homeViewModel =
+            ViewModelProvider(this).get(HomeViewModel:: class.java)
+
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textProfile
-        profileViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textHome
+        homeViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
 
@@ -42,4 +39,5 @@ class ProfileFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
+
 }
