+++ app/src/main/java/com/example/tp3_parcial_belgrano_grupo4/ui/views/viewmodels/ProfileViewModel.kt
@@ -1,12 +1,12 @@
-package com.example.tp3_parcial_belgrano_grupo4.ui.views.home
+package com.example.tp3_parcial_belgrano_grupo4.ui.views.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class ProfileViewModel : ViewModel() {
     private val _text = MutableLiveData<String>().apply {
-        value = "This is Home Fragment"
+        value = "This is Profile Fragment"
     }
     val text: LiveData<String> = _text
 }
