+++ app/src/main/java/com/example/tp3_parcial_belgrano_grupo4/ui/views/viewmodels/SettingViewModel.kt
@@ -1,12 +1,12 @@
-package com.example.tp3_parcial_belgrano_grupo4.ui.views.profile
+package com.example.tp3_parcial_belgrano_grupo4.ui.views.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ProfileViewModel : ViewModel() {
+class SettingViewModel : ViewModel() {
     private val _text = MutableLiveData<String>().apply {
-        value = "This is Profile Fragment"
+        value = "This is Setting Fragment"
     }
     val text: LiveData<String> = _text
 }
