+++ app/src/main/res/layout/activity_main.xml
@@ -8,8 +8,6 @@ android:layout_height="match_parent"
 android:orientation="vertical"
 tools:openDrawer="start">
 
-
-<!-- Contenedor del NavHostFragment -->
 <androidx.fragment.app.FragmentContainerView
     android:id="@+id/nav_host_fragment"
     android:name="androidx.navigation.fragment.NavHostFragment"
