+++ app/src/main/res/layout/fragment_setting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.views.setting.SettingFragment">
+    tools:context=".ui.views.fragments.SettingFragment">
 
     <TextView
         android:id="@+id/textSetting"
