+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,8 +6,8 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.example.tp3_parcial_belgrano_grupo4.ui.views.home.HomeFragment"
-        android:label="Home"
+        android:name="com.example.tp3_parcial_belgrano_grupo4.ui.views.fragments.HomeFragment"
+        android:label="@string/title_home"
         tools:layout="@layout/fragment_home">
         <action
             android:id="@+id/action_nav_home_to_nav_profile"
@@ -19,15 +19,15 @@
 
     <fragment
         android:id="@+id/nav_profile"
-        android:name="com.example.tp3_parcial_belgrano_grupo4.ui.views.profile.ProfileFragment"
-        android:label="Profile"
+        android:name="com.example.tp3_parcial_belgrano_grupo4.ui.views.fragments.ProfileFragment"
+        android:label="@string/title_profile"
         tools:layout="@layout/fragment_profile"/>
 
 
     <fragment
         android:id="@+id/nav_setting"
-        android:name="com.example.tp3_parcial_belgrano_grupo4.ui.views.setting.SettingFragment"
-        android:label="Setting"
+        android:name="com.example.tp3_parcial_belgrano_grupo4.ui.views.fragments.SettingFragment"
+        android:label="@string/title_setting"
         tools:layout="@layout/fragment_setting"/>
 
 </navigation>
