+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,9 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+
+    <color name="colorPrimary">#FFFFFF</color> <!-- Color de la barra de acción -->
+    <color name="colorPrimaryVariant">#FFC107</color> <!-- Color de acento -->
+    <color name="colorOnPrimary">#000000</color>
+
 </resources>
