+++ app/src/main/res/values/strings.xml
@@ -3,17 +3,13 @@
     <string name="baseUrl">https://dog.ceo/api/breed/</string>
 
     //Navigation Drawer
-    <string name="title_profile">Profile</string>
-    <string name="title_setting">Setting</string>
+    <string name="title_profile">Perfil</string>
+    <string name="title_setting">Configuracion</string>
     <string name="title_home">Home</string>
     <string name="title_activity_menu_drawer">menu_drawer</string>
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="nav_header_title">Martin</string>
-    <string name="nav_header_subtitle">Hola</string>
     <string name="nav_header_desc">Navigation Image</string>
     <string name="action_settings">Settings</string>
     <string name="image">ImageView</string>
-
-
 </resources>
