+++ app/src/main/res/values/themes.xml
@@ -13,8 +13,11 @@
     </style>
 
     <style name="Theme.TP3_PARCIAL_BELGRANO_GRUPO4.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+
     <style name="Theme.TP3_PARCIAL_BELGRANO_GRUPO4.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
-    <style name="AppTheme" parent="Theme.MaterialComponents.Light.DarkActionBar"/>
+
+    <style name="AppTheme" parent="Theme.MaterialComponents.Light.DarkActionBar" />
+
     <style name="Theme.DrawerMenu.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
